Control Tab Refactor
